requirements.txt
streamlit==0.88.0
mysql-connector-python==8.0.26
